Readd default IWebProxy for all non-Core platforms
For Android/iOS we need to set the HttpClientHandler.Proxy property to
the WebRequest.DefaultWebProxy. There is an issue in the Mono framework
which means it's inconsistent with the behaviour of the .NET Framework:

 In NetFx HttpClientHandler.Proxy is set to the WR.DefaultWebProxy if no
 proxy is explicitly set.
 In Mono the HttpClientHandler.Proxy property is NOT defaulted to
 WR.DefaultWebProxy.

Because .NET Standard 1.1 (or any version less than 2.0) does not
include the default IWebRequest implmentation (WebRequest) we introduce
the IWebProxyProvider platform abstraction.
In all platforms except CoreCLR (.NET Standard 1.1) we return the
WR.DefaultWebProxy. For the .NET Standard target platform we return null
since there's not much that can be done (the advice is to set proxies in
the network stack/OS level).